LayoutTests/crypto/subtle/rsa-oaep-import-key-unwrap-jwk-oct-key-expected.txt

Test unwrapping a JWK oct key with RSA-OAEP using an imported key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS unwrappedKey.kty is 'oct'
PASS Base64URL.parse(unwrappedKey.k) is rawKey
PASS unwrappedKey.alg is 'A128CBC'
PASS unwrappedKey.key_ops is ['decrypt', 'encrypt']
PASS unwrappedKey.ext is true
PASS successfullyParsed is true

TEST COMPLETE
